fix puppy build with !jmx
