+++ cmd/agent/api/agent/agent.go
@@ -12,23 +12,17 @@ import (
 	"encoding/json"
 	"fmt"
 	"net/http"
-	"strconv"
-	"time"
 
 	log "github.com/cihub/seelog"
 
 	"github.com/DataDog/datadog-agent/cmd/agent/common"
 	"github.com/DataDog/datadog-agent/cmd/agent/common/signals"
 	apiutil "github.com/DataDog/datadog-agent/pkg/api/util"
-	"github.com/DataDog/datadog-agent/pkg/collector/check"
-	"github.com/DataDog/datadog-agent/pkg/collector/corechecks/embed"
 	"github.com/DataDog/datadog-agent/pkg/flare"
 	"github.com/DataDog/datadog-agent/pkg/status"
 	"github.com/DataDog/datadog-agent/pkg/util"
 	"github.com/DataDog/datadog-agent/pkg/version"
 	"github.com/gorilla/mux"
-
-	yaml "gopkg.in/yaml.v2"
 )
 
 // SetupHandlers adds the specific handlers for /agent endpoints
@@ -119,102 +113,6 @@ func componentStatusHandler(w http.ResponseWriter, r *http.Request) {
 		log.Errorf("%s", err.Error())
 		http.Error(w, err.Error(), 404)
 	}
-}
-
-func getJMXConfigs(w http.ResponseWriter, r *http.Request) {
-	var err error
-
-	if err := apiutil.Validate(w, r); err != nil {
-		return
-	}
-
-	var ts int
-	queries := r.URL.Query()
-	if timestamps, ok := queries["timestamp"]; ok {
-		ts, _ = strconv.Atoi(timestamps[0])
-	}
-
-	if int64(ts) > embed.JMXConfigCache.GetModified() {
-		w.WriteHeader(http.StatusNoContent)
-		return
-	}
-
-	w.Header().Set("Content-Type", "application/json")
-	log.Debugf("Getting latest JMX Configs as of: %#v", ts)
-
-	j := map[string]interface{}{}
-	configs := map[string]check.ConfigJSONMap{}
-
-	configItems := embed.JMXConfigCache.Items()
-	for name, config := range configItems {
-		m, ok := config.(map[string]interface{})
-		if !ok {
-			err = fmt.Errorf("wrong type in cache")
-			log.Errorf("%s", err.Error())
-			http.Error(w, err.Error(), 500)
-			return
-		}
-
-		cfg, ok := m["config"].(check.Config)
-		if !ok {
-			err = fmt.Errorf("wrong type for config")
-			log.Errorf("%s", err.Error())
-			http.Error(w, err.Error(), 500)
-			return
-		}
-
-		var rawInitConfig check.ConfigRawMap
-		err = yaml.Unmarshal(cfg.InitConfig, &rawInitConfig)
-		if err != nil {
-			log.Errorf("unable to parse JMX configuration: %s", err)
-			http.Error(w, err.Error(), 500)
-			return
-		}
-
-		c := map[string]interface{}{}
-		c["init_config"] = util.GetJSONSerializableMap(rawInitConfig)
-		instances := []check.ConfigJSONMap{}
-		for _, instance := range cfg.Instances {
-			var rawInstanceConfig check.ConfigJSONMap
-			err = yaml.Unmarshal(instance, &rawInstanceConfig)
-			if err != nil {
-				log.Errorf("unable to parse JMX configuration: %s", err)
-				http.Error(w, err.Error(), 500)
-				return
-			}
-			instances = append(instances, util.GetJSONSerializableMap(rawInstanceConfig).(check.ConfigJSONMap))
-		}
-
-		c["instances"] = instances
-		configs[name] = c
-
-	}
-	j["configs"] = configs
-	j["timestamp"] = time.Now().Unix()
-	jsonPayload, err := json.Marshal(util.GetJSONSerializableMap(j))
-	if err != nil {
-		log.Errorf("unable to parse JMX configuration: %s", err)
-		http.Error(w, err.Error(), 500)
-		return
-	}
-	w.Write(jsonPayload)
-}
-
-func setJMXStatus(w http.ResponseWriter, r *http.Request) {
-	if err := apiutil.Validate(w, r); err != nil {
-		return
-	}
-
-	decoder := json.NewDecoder(r.Body)
-
-	var jmxStatus status.JMXStatus
-	err := decoder.Decode(&jmxStatus)
-	if err != nil {
-		log.Errorf("unable to parse jmx status: %s", err)
-		http.Error(w, err.Error(), 500)
-	}
-
-	status.SetJMXStatus(jmxStatus)
 }
 
 func getStatus(w http.ResponseWriter, r *http.Request) {
